Move protobuf contract and mappers to same folder as the integration event to improve overview and cohesiveness
