+++ source/business-workflow/source/Energinet.DataHub.MarketRoles.Infrastructure/Integration/IntegrationEventDispatching/EnergySupplierChange/EnergySupplierChangedIntegrationMapper.cs
@@ -13,22 +13,23 @@
 // limitations under the License.
 
 using System;
+using Energinet.DataHub.MarketRoles.Domain.MeteringPoints.Events;
 using Energinet.DataHub.MarketRoles.Infrastructure.Transport.Protobuf;
 using Energinet.DataHub.MarketRoles.IntegrationEventContracts;
 using Google.Protobuf;
 
-namespace Energinet.DataHub.MarketRoles.Infrastructure.Integration.IntegrationEventDispatching.ChangeOfSupplier.Mappers
+namespace Energinet.DataHub.MarketRoles.Infrastructure.Integration.IntegrationEventDispatching.EnergySupplierChange
 {
     public class EnergySupplierChangedIntegrationMapper : ProtobufOutboundMapper<EnergySupplierChangedIntegrationEvent>
     {
         protected override IMessage Convert(EnergySupplierChangedIntegrationEvent obj)
         {
             if (obj == null) throw new ArgumentException(null, nameof(obj));
-            return new EnergySupplierChanged
+            return new IntegrationEventContracts.EnergySupplierChanged()
             {
-                MeteringPointId = obj.GsrnNumber,
-                EnergySupplierGLN = obj.EnergySupplierGln,
                 EffectiveDate = obj.StartOfSupplyDate.ToString(),
+                GsrnNumber = obj.GsrnNumber,
+                EnergySupplierGln = obj.EnergySupplierGln,
             };
         }
     }
